프로필 상세 페이지 반응형 작업
