+++ package.json
@@ -15,7 +15,7 @@
     "react-easy-crop": "^5.0.4",
     "react-infinite-scroll-component": "^6.1.0",
     "react-intersection-observer": "^9.5.3",
-    "react-loading-skeleton": "^3.3.1",
+    "react-loading-skeleton": "^3.4.0",
     "react-router-dom": "^6.18.0",
     "react-scripts": "5.0.1",
     "recoil": "^0.7.7",
@@ -32,7 +32,6 @@
     "eject": "react-scripts eject",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build"
-
   },
   "eslintConfig": {
     "extends": [
