+++ src/Components/Profile/PostCount.jsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from 'react';
 
 import PropTypes from 'prop-types'; // npm install prop-types 설치 필요
-import styled from 'styled-components';
 
 import { userPostList } from '../../API/Post/PostAPI';
 import { productList } from '../../API/Product/ProductAPI';
+import { PostCountWrap } from './ProfileStyle';
 
 export default function PostCount({ accountName, userType }) {
   const [postCount, setPostCount] = useState(0);
@@ -36,18 +36,3 @@ PostCount.propTypes = {
   accountName: PropTypes.string.isRequired,
   userType: PropTypes.string.isRequired,
 };
-
-const PostCountWrap = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-
-  p {
-    font-size: 18px; // 여기에서 텍스트 사이즈를 10px로 변경합니다.
-  }
-
-  p:last-child {
-    font-size: 10px; // 첫 번째 p 태그 (게시글 수 숫자)는 기존대로 유지합니다.
-    color: #767676;
-  }
-`;
