+++ src/Pages/Home/HomeSearch.jsx
@@ -19,4 +19,3 @@ export default function HomeFeed() {
     </UserSearchHome>
   );
 }
-
