+++ src/Pages/Profile/MyProfile.jsx
@@ -2,7 +2,7 @@
   설명: 사용자 accountname의 프로필 페이지(내 페이지)
   작성자: 이해지
   최초 작성 날짜: 2023.10.29
-  마지막 수정 날까: 2023.12.08
+  마지막 수정 날까: 2024.02.03
 */
 
 import React, { useState, useEffect } from 'react';
@@ -18,10 +18,10 @@ import { Container } from '../../Components/Common/Container';
 import userToken from '../../Recoil/userTokenAtom'; //파일경로 변경 완료
 
 import Footer from '../../Components/Common/Footer';
-import styled from 'styled-components';
 import Header from '../../Components/Common/Header/Header';
 
 import { getMyProfileData } from '../../API/Profile/ProfileAPI';
+import { ProfileWrap, ProfileTop, Btns, MyProfileBtn } from './ProfileStyle';
 
 // 프로필보기
 function MyProfile() {
@@ -72,29 +72,29 @@ function MyProfile() {
       <Header type='moreKebab' />
       <ProfileWrap>
         <section>
-          <HiddenH1>내 프로필</HiddenH1>
+          <h1 className='a11y-hidden'>내 프로필</h1>
           <ProfileTop>
             <section>
               <ProfileDetail userInfoData={userInfoData} />
               <Btns>
-                <button
+                <MyProfileBtn
                   type='button'
                   onClick={() => {
                     navigate('/profile/edit');
                   }}
                 >
                   프로필 수정
-                </button>
+                </MyProfileBtn>
                 {/* 일반 계정일 경우 상품등록 버튼 제거 */}
                 {userType === 'organization' ? (
-                  <button
+                  <MyProfileBtn
                     type='button'
                     onClick={() => {
                       navigate('/product');
                     }}
                   >
                     행사 등록
-                  </button>
+                  </MyProfileBtn>
                 ) : null}
               </Btns>
             </section>
@@ -103,7 +103,6 @@ function MyProfile() {
             <ProfileDetailProduct userInfoData={userInfoData} reFetchInfo={UserInfo} />
           ) : null}
           <ProfileDetailPost accountName={profileAccountname} />
-
           <Footer />
         </section>
       </ProfileWrap>
@@ -112,43 +111,3 @@ function MyProfile() {
 }
 
 export default MyProfile;
-
-const a11yHidden = `
-  clip: rect(1px, 1px, 1px, 1px);
-  clip-path: inset(50%);
-  width: 1px;
-	height: 1px;
-	margin: -1px;
-  overflow: hidden;
-	padding: 0;
-	position: absolute;
-`;
-
-const HiddenH1 = styled.h1`
-  ${a11yHidden}
-`;
-const ProfileWrap = styled.div`
-  background-color: #fff;
-`;
-const ProfileTop = styled.div`
-  position: relative;
-  margin-top: 48px;
-`;
-
-const Btns = styled.div`
-  position: absolute;
-  top: 128px;
-  right: 16px;
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-  button {
-    padding: 8px 23px;
-    border: 1px solid var(--DBDBDB, #dbdbdb);
-    border-radius: 30px;
-    font-size: 14px;
-    font-weight: 700;
-
-    color: #767676;
-  }
-`;
